Handle ErrorKind::Interrupted in ReadMany and Eof by retrying
Convenience methods for readers in std usually handly ErrorKind::Interrupted
internally by retrying instead of returning the error to the user. These
errors can for example happen when an unix process recieves a signal.

This commits adds ErrorKind::Interrupted handling to ReadMany::read_many and
Eof::eof. For consistency, it is also added to <Eof as Read>::read. This is
a breaking change.

Fixes #1